+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+# [1.2.0](https://github.com/johnnymillergh/javascript-playground/compare/v1.1.1...v1.2.0) (2020-08-28)
+
+
+### Features
+
+* **LeetCode:** add solution to 13 roman to integer ([de72a87](https://github.com/johnnymillergh/javascript-playground/commit/de72a8747e52b74edafb16476b1d2f940ba64f56))
+* **LeetCode:** add solution to 14 longest common prefix ([4811774](https://github.com/johnnymillergh/javascript-playground/commit/4811774ee1fa919b843737ccb0ff3901004a26d7))
+
+
+### Performance Improvements
+
+* **lint-staged:** update configuration ([39b1447](https://github.com/johnnymillergh/javascript-playground/commit/39b1447c0a0a470a5655fb8d11a69f5b1a22cfd3))
+* **Node.js:** update Node.js and NPM version constraints ([0f117b4](https://github.com/johnnymillergh/javascript-playground/commit/0f117b4ae2b6b97f987c0a89ed336c4c82071084))
+
+
+### Tests
+
+* **LeetCode:** add unit tests for solution to 13 roman to integer ([efca00e](https://github.com/johnnymillergh/javascript-playground/commit/efca00ed2a4dcd7788ff7879d317ee5f89ac39e4))
+
+
+### BREAKING CHANGES
+
+* **lint-staged:** enable auto EsLint check before commit
+* **LeetCode:** inline unit tests
+* **LeetCode:** integrate LeetCode solution in JavaScript
+
+
+
 ## [1.1.1](https://github.com/johnnymillergh/javascript-playground/compare/v1.1.0...v1.1.1) (2020-08-26)